feat: Se añade vista de resultado
