+++ package.json
@@ -29,6 +29,7 @@
     "clsx": "^2.1.1",
     "embla-carousel-autoplay": "^8.1.7",
     "embla-carousel-react": "^8.1.7",
+    "html-react-parser": "^5.1.12",
     "jsdom": "24.1.0",
     "lucide-react": "^0.414.0",
     "next": "14.2.5",
